Fix "AUTOSAR C++14 A4-7-1" rule findings
This rule states: "An integer expression shall not lead to data loss.".

GIT_ORIGIN_SPP_REV_ID: 26daf77d2dcf0ea9ef287014f5ba601b3e3678ac

+++ score/message_passing/qnx_dispatch/qnx_dispatch_engine.h
@@ -213,14 +213,14 @@ class QnxDispatchEngine final : public ISharedResourceEngine
     }
 
   private:
-    enum class PulseEvent : std::uint8_t
+    enum class PulseEvent : std::int32_t
     {
         QUIT,
         TIMER
     };
 
     void SendPulseEvent(const PulseEvent pulse_event) noexcept;
-    void ProcessPulseEvent(const std::uint8_t pulse_event) noexcept;
+    void ProcessPulseEvent(const std::int32_t pulse_event) noexcept;
     void ProcessCleanup(const void* const owner) noexcept;
     void RunOnThread() noexcept;
